requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cssselect==1.3.0
dataclasses-json==0.6.7
distro==1.9.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
Flask==3.1.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
itsdangerous==2.2.0
jieba3k==0.35.1
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-openai==0.3.12
langchain-text-splitters==0.3.8
langsmith==0.3.25
lxml==5.3.2
lxml_html_clean==0.4.1
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.3.2
mypy-extensions==1.0.0
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.2.4
openai==1.71.0
orjson==3.10.16
packaging==24.2
pillow==11.1.0
propcache==0.3.1
pydantic==2.11.2
pydantic-settings==2.8.1
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
tenacity==9.1.2
tiktoken==0.9.0
tinysegmenter==0.3
tldextract==5.2.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
Werkzeug==3.1.3
yarl==1.19.0
zstandard==0.23.0
